+++ src/pages/api-forms/index.tsx
@@ -1,0 +1,5 @@
+import { ApiForms } from "@/features/ApiForms";
+
+export default function ApiFormsPage() {
+  return <ApiForms />;
+}
