+++ src/types/response/index.ts
@@ -1,0 +1,3 @@
+export type ResponseData = {
+  error?: string;
+}
